Use dual write handler for role in role migrator
