+++ rbac/migration_tool/migrate.py
@@ -16,86 +16,26 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
 import logging
-from typing import Iterable
 
 from django.conf import settings
 from django.db import transaction
-from kessel.relations.v1beta1 import common_pb2
 from management.group.relation_api_dual_write_group_handler import RelationApiDualWriteGroupHandler
-from management.models import Group, Workspace
+from management.models import Group
 from management.principal.model import Principal
 from management.relation_replicator.logging_replicator import LoggingReplicator
 from management.relation_replicator.outbox_replicator import OutboxReplicator
 from management.relation_replicator.relation_replicator import (
-    PartitionKey,
     RelationReplicator,
-    ReplicationEvent,
     ReplicationEventType,
 )
 from management.relation_replicator.relations_api_replicator import RelationsApiReplicator
-from management.role.model import BindingMapping, Role
-from migration_tool.models import V2rolebinding
-from migration_tool.sharedSystemRolesReplicatedRoleBindings import v1_role_to_v2_bindings
-from migration_tool.utils import create_relationship
+from management.role.model import Role
+from management.role.relation_api_dual_write_handler import RelationApiDualWriteHandler
 
 from api.cross_access.relation_api_dual_write_cross_access_handler import RelationApiDualWriteCrossAccessHandler
 from api.models import CrossAccountRequest, Tenant
 
-
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
-
-
-def get_kessel_relation_tuples(
-    v2_role_bindings: Iterable[V2rolebinding],
-    default_workspace: Workspace,
-) -> list[common_pb2.Relationship]:
-    """Generate a set of relationships and BindingMappings for the given set of v2 role bindings."""
-    relationships: list[common_pb2.Relationship] = list()
-
-    for v2_role_binding in v2_role_bindings:
-        relationships.extend(v2_role_binding.as_tuples())
-
-        bound_resource = v2_role_binding.resource
-
-        # Is this a workspace binding, but not to the root workspace?
-        # If so, ensure this workspace is a child of the root workspace.
-        # All other resource-resource or resource-workspace relations
-        # which may be implied or necessary are intentionally ignored.
-        # These should come from the apps that own the resource.
-        if bound_resource.resource_type == ("rbac", "workspace") and not bound_resource.resource_id == str(
-            default_workspace.id
-        ):
-            # This is not strictly necessary here and the relation may be a duplicate.
-            # Once we have more Workspace API / Inventory Group migration progress,
-            # this block can and probably should be removed.
-            # One of those APIs will add it themselves.
-            relationships.append(
-                create_relationship(
-                    bound_resource.resource_type,
-                    bound_resource.resource_id,
-                    ("rbac", "workspace"),
-                    str(default_workspace.id),
-                    "parent",
-                )
-            )
-
-    return relationships
-
-
-def migrate_role(
-    role: Role,
-    default_workspace: Workspace,
-    current_bindings: Iterable[BindingMapping] = [],
-) -> tuple[list[common_pb2.Relationship], list[BindingMapping]]:
-    """
-    Migrate a role from v1 to v2, returning the tuples and mappings.
-
-    The mappings are returned so that we can reconstitute the corresponding tuples for a given role.
-    This is needed so we can remove those tuples when the role changes if needed.
-    """
-    v2_role_bindings = v1_role_to_v2_bindings(role, default_workspace, current_bindings)
-    relationships = get_kessel_relation_tuples([m.get_role_binding() for m in v2_role_bindings], default_workspace)
-    return relationships, v2_role_bindings
 
 
 def migrate_groups_for_tenant(tenant: Tenant, replicator: RelationReplicator):
@@ -133,8 +73,6 @@ def migrate_groups_for_tenant(tenant: Tenant, replicator: RelationReplicator):
 
 def migrate_roles_for_tenant(tenant, exclude_apps, replicator):
     """Migrate all roles for a given tenant."""
-    default_workspace = Workspace.objects.get(type=Workspace.Types.DEFAULT, tenant=tenant)
-
     roles = tenant.role_set.all()
     if exclude_apps:
         roles = roles.exclude(access__permission__application__in=exclude_apps)
@@ -142,24 +80,13 @@ def migrate_roles_for_tenant(tenant, exclude_apps, replicator):
     for role in roles:
         logger.info(f"Migrating role: {role.name} with UUID {role.uuid}.")
 
-        tuples, mappings = migrate_role(role, default_workspace)
-
-        # Conflicts are not ignored in order to prevent this from
-        # accidentally running concurrently with dual-writes.
-        # If migration should be rerun, then the bindings table should be dropped.
-        # If changing this to allow updates,
-        # always ensure writes are paused before running.
-        # This must always be the case, but this should at least start failing you if you forget.
-        BindingMapping.objects.bulk_create(mappings, ignore_conflicts=False)
-
-        replicator.replicate(
-            ReplicationEvent(
-                event_type=ReplicationEventType.MIGRATE_CUSTOM_ROLE,
-                info={"role_uuid": str(role.uuid)},
-                partition_key=PartitionKey.byEnvironment(),
-                add=tuples,
+        with transaction.atomic():
+            role = Role.objects.select_for_update().get(pk=role.pk)
+            dual_write_handler = RelationApiDualWriteHandler(
+                role, ReplicationEventType.MIGRATE_CUSTOM_ROLE, replicator
             )
-        )
+
+            dual_write_handler.replicate_new_or_updated_role(role)
 
         logger.info(f"Migration completed for role: {role.name} with UUID {role.uuid}.")
     logger.info(f"Migrated {roles.count()} roles for tenant: {tenant.org_id}")
